requirements.txt
Django==1.4.3
South==0.7.6
argparse==1.2.1
distribute==0.6.24
django-tastypie==0.9.12-alpha
ipython==0.13.1
mimeparse==0.1.3
pysqlite==2.6.3
python-dateutil==1.5
wsgiref==0.1.2
